Add test for fetching the countries on the SupportedCountriesViewModel
SupportedCountriesViewModel has been added to manage the business logic and
interact with CountryRepository which will obtain the data
that will be shown on the SupportedCountriesFragment.
A test has been created for validating the case of obtaining
and exposing the countries to the SupportedCountriesFragment.
The arch-core libraries had been added to help with testing LiveData within the
SupportedCountriesViewModel. The class LiveDataTestUtil has been copied from
https://github.com/android/architecture-samples/blob/master/app/src/test/java/com/example/android/architecture/blueprints/todoapp/LiveDataTestUtil.kt
to help with it too.
The mockito library has been added to help with creating mocks
of the CountryRepository on the SupportedCountriesViewModelTest.