Add integration with Fixer API
Added Retrofit library for handling network requests.
Added Moshi library to parse the JSON responses.
Added a logging interceptor for easier debugging.
Used flow for managing the response from the requests.
Changed the functions on the repository, the data source and the
view model to suspend functions to handle running code on a new thread.
Added the internet permissions on the AndroidManifest.
Added a security configuration because on the newer versions of android
HTTP requests are disable by default and the API only works with
those type of requests.